+++ splinter/driver/webdriver/firefox.py
@@ -33,13 +33,13 @@ class WebDriver(BaseWebDriver):
 
 class WebDriverElement(BaseWebDriverElement):
 
-    def mouseover(self):
+    def mouse_over(self):
         """
         Firefox doesn't support mouseover.
         """
         raise NotImplementedError("Firefox doesn't support mouse over")
 
-    def mouseout(self):
+    def mouse_out(self):
         """
         Firefox doesn't support mouseout.
         """
@@ -50,3 +50,18 @@ class WebDriverElement(BaseWebDriverElement):
         Firefox doesn't support doubleclick.
         """
         raise NotImplementedError("Firefox doesn't support doubleclick")
+
+    def right_click(self):
+        """
+        Firefox doesn't support right click'
+        """
+        raise NotImplementedError("Firefox doesn't support right click")
+
+    def drag_and_drop(self, droppable):
+        """
+        Firefox doesn't support drag and drop
+        """
+        raise NotImplementedError("Firefox doesn't support drag an drop")
+
+    mouseover = mouse_over
+    mouseout = mouse_out