+++ tests/test_zopetestbrowser.py
@@ -38,10 +38,13 @@ class ZopeTestBrowserDriverTest(BaseBrowserTests, unittest.TestCase):
         assert_equals(EXAMPLE_APP, browser.url)
         browser.quit()
 
-    @raises(NotImplementedError)
     def test_cant_switch_to_frame(self):
         "zope.testbrowser should not be able to switch to frames"
-        self.browser.get_iframe('frame_123')
+        try:
+            self.browser.get_iframe('frame_123')
+            self.fail()
+        except NotImplementedError as e:
+            assert_equals("zope.testbrowser does not support frames", e.args[0])
 
     @raises(NotImplementedError)
     def test_simple_type(self):
@@ -56,9 +59,9 @@ class ZopeTestBrowserDriverTest(BaseBrowserTests, unittest.TestCase):
     @raises(NotImplementedError)
     def test_cant_mouseover(self):
         "zope.testbrowser should not be able to put the mouse over the element"
-        self.browser.find_by_css('#visible').first.mouseover()
+        self.browser.find_by_css('#visible').first.mouse_over()
 
     @raises(NotImplementedError)
     def test_cant_mouseout(self):
         "zope.testbrowser should not be able to mouse out of an element"
-        self.browser.find_by_css('#visible').first.mouseout()
+        self.browser.find_by_css('#visible').first.mouse_out()